disk: Cast fsblkcnt_t to double instead of float

+++ components/disk.c
@@ -29,7 +29,7 @@ disk_perc(const char *path)
 	}
 
 	return bprintf("%d", (int)(100 *
-	               (1.0f - ((float)fs.f_bavail / (float)fs.f_blocks))));
+	               (1 - ((double)fs.f_bavail / (double)fs.f_blocks))));
 }
 
 const char *
